chore: modular MoreMenu
